only handle volumes created by plugin
On Create() we set a plugin unique, so called aux property, and when
listing volumes, we filter by resources having this property.

This breaks compatibility with earlier versions of this plugin. A user
can set the aux property manually via the linstor client. I guess at
this early stage of the plugin this is okay.

This also adds two consts for the aux key and value. Note that the key
has to start with "Aux/" in order to be a key that can be set by the
user.

Closes #3