handle sizes with units
This adds units to the Size option and does the usual checks. If not set
at all, set to 100MB (default from the Python based plugin). I did not
care to make MB the default if no unit is specified. This is a trade-off
I am willing to take. The plugin should be "sufficiently similar" to the
Python one, but it does not need to be a pixel perfect copy.

This implements one part of #4.