let LINSTOR do mkfs, add FSOpts
The Python based plugin has "fsopts", where one can specify additional
parameters. The problem is that we did not have that part under control
by using the k8s mounter utils. The good thing is that LINSTOR itself
can create a file system on top of the device.

- use LINSTOR to create the FS. This is a good idea in general, LINSTOR
only supports xfs and ext4, but it already sets some mkfs options
accordingly.
- add FSOpts for additional options.
- use only the Mount()er part from k8s.

Again, this does not make the plugin pixel perfect to the Python one,
but IMO with the previous commit, we can now:

Close #4